parse.go: Use reflect to get a pointer to NodeList
A pointer is used to append Nodes to a NodeList. When a Node of type
NodeSection is found, any Nodes found after the section are appended to the
NodeList field of that section. Reflection is used to get a pointer to the
"NodeList" field since Interfaces only expose methods.

It works, but I am not sure if it is idiomatic Go since it feels like kind of a
kludge.