Major Parser Change: NodeParagraph is a container
Previously, NodeParagraph was only simple text. If inline markup was
defined in a paragraph, then the structure for a document such as:

    first *emphasized* text

    second *emphasized* text

would be:

    NodeParagraph "first"
    NodeEmphasis "emphasized"
    NodePargraph "text"
    NodeParagraph "second"
    NodeEmphasis "emphasized"
    NodePargraph "text"

This clearly will not scale and easily becomes confusing.

This change makes NodeParagraph a container of nodes using a NodeList. A
new node type is added: NodeText, which becomes the regular text. So
now, the structure for a line above would look something like,

    NodeParagraph
        NodeText "first"
        NodeEmphasis "emphasized"
        NodeText"text"
    NodeParagraph
        NodeText "second"
        NodeEmphasis "emphasized"
        NodeText"text"

Unfortunately, this change will break many tests... :(