Remove ID from Node interface
In an effort to stream line *-nodes.json tests, the ID field has been
removed. The ID field made it easier to find a specific node when a test
failed, but the time needed to support the ID field is not worth it.

Whenever a node changed, then the ID must be changed. A script was made
to help with auto renumbering...