Switch logging library to go-kit/log
This new logging library gives more info and requires less typing. The
"logger.WithFields(logger.Fields{})" was getting tiresome to type and
was cluttering the code. The go-kit logger provides all of that but
with an ergonomic api.

While the go-kit logger does provide for some great primitives, I am
still not certain if it will allow me to put something together to
easily debug go-rst. I have therefore abstracted the logging into a few
function calls that will make it easier to change logging libraries in
the future if needed.  Hopefully this fourth logging library switch will
be the last.