Reorder tests
I tried to reorder the tests into a sensible top down order for
implementation. In theory, the tests can be implemented in order
starting at 00, 01, 02, ...

Still, there will be tests in 02 that will have some dependency on 05
for example. When this happens you can either skip the test in 02 or
start implementing the 05 dependency. The reStructuredText spec does not
give any details on order of implementation.