testutil: fix bad logic bug in JsonDiff()
The parsed/lexed items and the expected items were not being checked
correctly due to a logic error. 10 tests were found to be broken or
incorrect, tens more were modified to explicitly set the startPosition.
Previously, it was acceptable to omit startPositon if it was 1.
StartPosition cannot be ommitted from the JSON output because is
initialized to 1, and thus not empty.