Fix duplicate test IDs
There were numerous tests that shared the same test "dot quad" id. This
made it confusing when filtering a single test id because there would be
multiple tests with the same id.

In order to fix this, I had to reorganize the testdata directory to make
it harder to duplicate IDs. I also removed the 01-good and 02-bad
directories that added unneeded depth to the testdata tree.

In the future I will be enforcing the test id uniqueness requirement
through testing and also auto-generating the Go tests. Auto generating
the Go tests for the testdata directory will help make it easier to
refactor and help ensure correctness.