Refactor of Token Buffer
To implement inline markup in section title, a larger token buffer is
needed so that it is possible to peek back far enough to get the initial
section title token. With inline markup, the opening and closing inline
markup tokens use up the four token "back" buffer, resulting in the
initial section title and section overline adornment tokens are no
longer in the buffer.