Last work
One test is failing. I stopped working on this because there is very
little interest in a restructuredtext parser for Go and this is not an
easy format to parse. So I am stopping work on this for the time being.

The main reason I had for wanting to write a restructuredtext parser for
go was to write my notes and website in restructuredtext, but that need
has passed.